chore: correcting create new module skeleton
